Extract await helper from robust-periodic-scheduler-tests for use with default scheduler too
