+++ test/rx/concurrency/test_default_scheduler.rb
@@ -1,21 +1,22 @@
 # Copyright (c) Microsoft Open Technologies, Inc. All rights reserved. See License.txt in the project root for license information.
 
 require 'test_helper'
+require 'rx/subscriptions/helpers/await_helpers'
 
-# DefaultScheduler creates new threads in which to run scheduled tasks; a short
-# sleep is necessary to allow the thread scheduler to yield to the other
-# threads.
 class TestDefaultScheduler < Minitest::Test
+  include AwaitHelpers
 
   def setup
     @scheduler = Rx::DefaultScheduler.instance
   end
 
+  INTERVAL = 0.05
+
   def test_schedule_with_state
     state = []
     task  = ->(_, s) { s << 1 }
     @scheduler.schedule_with_state(state, task)
-    sleep 0.001
+    await_array_length(state, 1, INTERVAL)
 
     assert_equal([1], state)
   end
@@ -24,15 +25,18 @@ class TestDefaultScheduler < Minitest::Test
     state = []
     task  = ->(_, s) { s << 1 }
     @scheduler.schedule_relative_with_state(state, 0.05, task)
-    sleep 0.1
+    await_array_length(state, 1, INTERVAL)
 
     assert_equal([1], state)
   end
 
   def test_default_schedule_runs_in_its_own_thread
+    state = []
     id = Thread.current.object_id
-    @scheduler.schedule -> { refute_equal(id, Thread.current.object_id) }
-    sleep 0.001
+    @scheduler.schedule -> { state << Thread.current.object_id }
+    await_array_length(state, 1, INTERVAL)
+
+    refute_equal([id], state)
   end
 
   def test_schedule_action_cancel
